Created Fortran namelist file for inputs
A namelist file is a simple text file that contains input parameters for a Fortran programme. This file contains input for module KWN-deformation. "inputs" is the name of the namelist, and each line contains the name and value. The line beginning with 'c' or '*' is for aesthetic or comment purposes only and will not be read. The / at the file's end signifies the end of the namelist. We can access these values using the READ command and namelist name 'inputs' in the Fortran programme.